+++ src/main/java/lk/server/cocktails/CocktailsApplication.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @EntityScan(basePackages = {
-        "lk.server.cocktails.db.cocktail.entities",
-        "lk.server.cocktails.db.ingredient.entities"
+        "lk.server.cocktails.features.cocktail.entities",
+        "lk.server.cocktails.features.ingredient.entities"
 })
 @EnableJpaRepositories(basePackages = {
-        "lk.server.cocktails.feachers.cocktail.repositories",
-        "lk.server.cocktails.feachers.ingredient.repositories"
+        "lk.server.cocktails.features.cocktail.repositories",
+        "lk.server.cocktails.features.ingredient.repositories"
 })
 public class CocktailsApplication {
