+++ src/main/java/lk/server/cocktails/configuration/files/FileManagerConfiguration.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class FileManagerConfiguration {
     @Bean
-    public FileManager getFileManager(){
+    public FileManager getFileManager() {
         return new FileManager();
     }
 }
