--- src/main/java/lk/server/cocktails/feachers/init/dto/IngredientFileStructure.java
@@ -1,15 +1,0 @@
-package lk.server.cocktails.feachers.init.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class IngredientFileStructure {
-
-    private String nameRU;
-    private String nameEN;
-
-}
