--- src/main/java/lk/server/cocktails/feachers/init/dto/help/SavedCocktailsDto2.java
@@ -1,16 +1,0 @@
-package lk.server.cocktails.feachers.init.dto.help;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SavedCocktailsDto2 {
-    private Long cocktailId;
-    private String cocktailName;
-    private List<Long> ingredientIds;
-}
