+++ src/main/java/lk/server/cocktails/features/ingredient/repositories/IngredientNameRepository.java
@@ -1,0 +1,10 @@
+package lk.server.cocktails.features.ingredient.repositories;
+
+import lk.server.cocktails.features.ingredient.entities.IngredientName;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IngredientNameRepository extends JpaRepository<IngredientName, Long> {
+
+}
