+++ src/main/java/lk/server/cocktails/features/init/dto/IngredientFileStructure.java
@@ -1,0 +1,15 @@
+package lk.server.cocktails.features.init.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class IngredientFileStructure {
+
+    private String nameRU;
+    private String nameEN;
+
+}
