+++ src/main/java/lk/server/cocktails/features/init/dto/help/SavedCocktailsDto.java
@@ -1,0 +1,16 @@
+package lk.server.cocktails.features.init.dto.help;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class SavedCocktailsDto {
+    private Long cocktailId;
+    private String cocktailName;
+    private List<String> ingredientNames;
+}
