+++ src/main/resources/application.properties
@@ -2,7 +2,6 @@
 # SERVER
 # ===============================
 server.port=9191
-
 # ===============================
 # DATABASE
 # ===============================
@@ -10,7 +9,6 @@ spring.datasource.driver-class-name=${DATABASE_DRIVER_CLASS_NAME}
 spring.datasource.url=jdbc:postgresql://${DATABASE_HOST}:${DATABASE_PORT}/${DATABASE_NAME}
 spring.datasource.username=${DATABASE_USERNAME}
 spring.datasource.password=${DATABASE_PASSWORD}
-
 # ===============================
 # JPA / HIBERNATE
 # ===============================
